publish,comment,relation,favorite

+++ database/model/base.go
@@ -7,7 +7,7 @@ import (
 )
 
 type BaseModel struct {
-	ID        uint           `gorm:"primaryKey" json:"id"`
+	ID        int64          `gorm:"primaryKey" json:"id"`
 	CreatedAt time.Time      `json:"createdAt"`
 	UpdatedAt time.Time      `json:"updatedAt"`
 	DeletedAt gorm.DeletedAt `gorm:"index" json:"-"`
